pyproject.toml
[project]
name = "process-intake-form"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai==1.58.1",
    "python-dotenv>=1.0.1",
    "structlog>=24.4.0",
]
